change url path of collections
